refactoring and troubleshoot, parked
