+++ packages/file-search/package.json
@@ -8,8 +8,7 @@
     "@theia/filesystem": "1.13.0",
     "@theia/process": "1.13.0",
     "@theia/workspace": "1.13.0",
-    "vscode-ripgrep": "^1.2.4",
-    "@theia/monaco": "1.13.0"
+    "vscode-ripgrep": "^1.2.4"
   },
   "publishConfig": {
     "access": "public"
